dom/media/webaudio/test/test_currentTime.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test AudioContext.currentTime</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("This test needs to wait a while for the AudioContext's timer to start.");
addLoadEvent(function() {
  var ac = new AudioContext();
  is(ac.currentTime, 0, "AudioContext.currentTime should be 0 initially");
  setTimeout(function() {
    ok(ac.currentTime > 0, "AudioContext.currentTime should have increased by now");
    SimpleTest.finish();
  }, 1000);
});

</script>
</pre>
</body>
</html>
